requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==25.4.0
audioread==3.0.1
av==16.0.1
blinker==1.9.0
cachetools==6.2.1
celluloid==0.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
decorator==5.2.1
deepspeed==0.12.4
Deprecated==1.2.18
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.3.0
fastapi==0.119.0
faster-whisper==1.2.0
ffmpy==0.6.3
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
gradio_client==0.7.1
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jaconv==0.4.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langsmith==0.4.37
lazy_loader==0.4
librosa==0.10.1
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.1
macholib==1.16.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mecab-python3==1.0.10
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
narwhals==2.8.0
networkx==3.4.2
ninja==1.13.0
numba==0.62.1
numpy==2.2.6
nvidia-ml-py==13.580.82
omegaconf==2.3.0
onnxruntime==1.23.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==10.4.0
pipdeptree==2.29.0
platformdirs==4.5.0
pooch==1.8.2
primePy==1.3
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
ptflops==0.7.1.2
py-cpuinfo==9.0.0
pyannote-audio==4.0.1
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydeck==0.9.1
pydomino @ file:///Users/taigamori/Works/tranosuke/pydomino
pydub==0.25.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pykakasi==2.3.0
pympi-ling==1.71
pynvml==13.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
resampy==0.4.2
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.44
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tabulate==0.8.10
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.12.0
toolz==1.0.0
torch==2.9.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.9.0
torchcodec==0.8.0
torchmetrics==1.8.2
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
webrtcvad==2.0.10
websockets==11.0.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
